third_party/blink/web_tests/external/wpt/webstorage/set.window-expected.txt

This is a testharness.js-based test.
PASS Setting property for key 9 on localStorage
PASS Setting property with toString for key 9 on localStorage
FAIL Setting property for key 9 on localStorage with data property on prototype assert_equals: expected "proto for Setting property for key 9 on localStorage with data property on prototype" but got "value for Setting property for key 9 on localStorage with data property on prototype"
FAIL Setting property for key 9 on localStorage with data property on prototype and existing item assert_equals: expected "proto for Setting property for key 9 on localStorage with data property on prototype and existing item" but got "existing for Setting property for key 9 on localStorage with data property on prototype and existing item"
FAIL Setting property for key 9 on localStorage with accessor property on prototype assert_equals: expected "proto getter for Setting property for key 9 on localStorage with accessor property on prototype" but got "value for Setting property for key 9 on localStorage with accessor property on prototype"
PASS Setting property for key x on localStorage
PASS Setting property with toString for key x on localStorage
FAIL Setting property for key x on localStorage with data property on prototype assert_equals: expected "proto for Setting property for key x on localStorage with data property on prototype" but got "value for Setting property for key x on localStorage with data property on prototype"
FAIL Setting property for key x on localStorage with data property on prototype and existing item assert_equals: expected "proto for Setting property for key x on localStorage with data property on prototype and existing item" but got "value for Setting property for key x on localStorage with data property on prototype"
FAIL Setting property for key x on localStorage with accessor property on prototype assert_equals: expected "proto getter for Setting property for key x on localStorage with accessor property on prototype" but got "value for Setting property for key x on localStorage with data property on prototype"
PASS Setting property for key 9 on sessionStorage
PASS Setting property with toString for key 9 on sessionStorage
FAIL Setting property for key 9 on sessionStorage with data property on prototype assert_equals: expected "proto for Setting property for key 9 on sessionStorage with data property on prototype" but got "value for Setting property for key 9 on sessionStorage with data property on prototype"
FAIL Setting property for key 9 on sessionStorage with data property on prototype and existing item assert_equals: expected "proto for Setting property for key 9 on sessionStorage with data property on prototype and existing item" but got "existing for Setting property for key 9 on sessionStorage with data property on prototype and existing item"
FAIL Setting property for key 9 on sessionStorage with accessor property on prototype assert_equals: expected "proto getter for Setting property for key 9 on sessionStorage with accessor property on prototype" but got "value for Setting property for key 9 on sessionStorage with accessor property on prototype"
PASS Setting property for key x on sessionStorage
PASS Setting property with toString for key x on sessionStorage
FAIL Setting property for key x on sessionStorage with data property on prototype assert_equals: expected "proto for Setting property for key x on sessionStorage with data property on prototype" but got "value for Setting property for key x on sessionStorage with data property on prototype"
FAIL Setting property for key x on sessionStorage with data property on prototype and existing item assert_equals: expected "proto for Setting property for key x on sessionStorage with data property on prototype and existing item" but got "value for Setting property for key x on sessionStorage with data property on prototype"
FAIL Setting property for key x on sessionStorage with accessor property on prototype assert_equals: expected "proto getter for Setting property for key x on sessionStorage with accessor property on prototype" but got "value for Setting property for key x on sessionStorage with data property on prototype"
Harness: the test ran to completion.
